[Chore] Decouple frontend tempates from API view
