+++ bucketlist/templates/bucketlists/base.html
@@ -1,0 +1,31 @@
+{% load staticfiles %}
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Just-Do-It</title>
+    <!-- google fonts -->
+    <link href="http://fonts.googleapis.com/css?family=Lobster&subset=latin,latin-ext" rel="stylesheet" type="text/css">
+    <!-- Material Design Icons -->
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!-- bootstrap core css -->
+    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
+    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap-theme.min.css">
+    <!-- inhouse design -->
+    <link rel="stylesheet" href="{% static 'css/bucketlist.css' %}">
+    <script src="static/js/main.js"></script>
+</head>
+<body>
+    <div class="page-header">
+        {% if user.is_authenticated %}
+        <p> Hey {{user.username}}</p>
+        {% else %}
+        <p> LogIn </p>
+        {% endif %}
+    </div>
+    <div class="content">
+        {% block content%}
+        {% endblock %}
+    </div>
+</body>
+</html>
